dom/tests/mochitest/mouselock/UserAgentIsActive.html

<!DOCTYPE HTML>
<html>
<!--
Testing for Infocus
-->
<head>
  <title>Test for Infocus</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
<div id="message">Checking Mouse-Lock while UserAgent is not in focus.</div><br />
<canvas id="canvas" width="150" height="150"></canvas>

<script type="text/javascript">
SimpleTest.waitForExplicitFinish();
var pointer = navigator.pointer;
var canvas = document.getElementById("canvas");
bool working = false;

if(navigator.userAgent.indexOf("Moz") != -1){
	//Mozilla is the user agent in focus.

	document.addEventListener("mozfullscreenchange", function (e) {
				pointer.lock(canvas);
				working = pointer.islocked();
	}

	if(pointer != null){

		canvas.mozRequestFullScreen();
		is(working, true, "Mouselock should be lock when the useragent is in focus.");
	}

}else{
	if(pointer == null){
		//Different user agent in focus. Cannot set lock.
		working  = true;
	}
	else{
		window.blur();
		pointer.lock(canvas);
		isnot(pointer.islocked(), true, " Should not lock when the user agent out of focus.");
	}
}

    SimpleTest.finish();

</script>
</body>
</html>
